refactor: library items now all importable under `library` package
