+++ alembic/env.py
@@ -19,7 +19,7 @@ from logging.config import fileConfig
 
 from alembic import context
 from moe.config import Config
-from moe.library import SABase
+from moe.library.lib_item import SABase
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
