+++ moe/library/__init__.py
@@ -1,5 +1,13 @@
 """Moe database/library functionality."""
 
-from sqlalchemy.orm import declarative_base
+from . import album, extra, lib_item, track
+from .album import *
+from .extra import *
+from .lib_item import *
+from .track import *
 
-SABase = declarative_base()
+__all__ = []
+__all__.extend(album.__all__)
+__all__.extend(extra.__all__)
+__all__.extend(lib_item.__all__)
+__all__.extend(track.__all__)
