+++ moe/plugins/moe_import/import_core.py
@@ -6,10 +6,7 @@ import pluggy
 
 import moe
 from moe import config
-from moe.library.album import Album
-from moe.library.extra import Extra
-from moe.library.lib_item import LibItem
-from moe.library.track import Track
+from moe.library import Album, Extra, LibItem, Track
 
 __all__ = ["import_album"]
