+++ tests/library/test_lib_item.py
@@ -5,9 +5,7 @@ import pytest
 
 import moe
 from moe.config import ExtraPlugin, MoeSession
-from moe.library.album import Album
-from moe.library.extra import Extra
-from moe.library.track import Track
+from moe.library import Album, Extra, Track
 from tests.conftest import album_factory, extra_factory, track_factory
