+++ tests/plugins/duplicate/test_dup_core.py
@@ -6,9 +6,7 @@ import pytest
 
 import moe
 from moe.config import ExtraPlugin, MoeSession
-from moe.library.album import Album
-from moe.library.extra import Extra
-from moe.library.track import Track
+from moe.library import Album, Extra, Track
 from moe.plugins.remove import remove_item
 from tests.conftest import album_factory, extra_factory, track_factory
