+++ tests/test_query.py
@@ -4,9 +4,7 @@
 import pytest
 
 import moe.query
-from moe.library.album import Album
-from moe.library.extra import Extra
-from moe.library.track import Track
+from moe.library import Album, Extra, Track
 from moe.query import QueryError, query
 from tests.conftest import album_factory, extra_factory, track_factory
